+++ grid2op/Runner/Runner.py
@@ -225,7 +225,9 @@ class Runner(object):
                  opponent_action_class=DontAct,
                  opponent_class=BaseOpponent,
                  opponent_init_budget=0,
-                 grid_layout=None):
+                 grid_layout=None,
+                 forbid_dispatch_off=False,
+                 ignore_min_up_down_times=True):
         """
         Initialize the Runner.
 
@@ -282,10 +284,18 @@ class Runner(object):
         voltagecontrolerClass: :class:`grid2op.VoltageControler.ControlVoltageFromFile`, optional
             The controler that will change the voltage setpoints of the generators.
 
+        forbid_dispatch_off: ``bool``
+            Whether or not you forbid to apply dispacthing action on turned off generator (default False)
+
+        ignore_min_up_down_times: ``bool``
+            Whether or not to ignore gen_min_uptime and gen_min_downtime when applying redispatching actions.
+            (default True)
+
         """
 
         self.name_env = name_env
-
+        self.ignore_min_up_down_times = ignore_min_up_down_times
+        self.forbid_dispatch_off = forbid_dispatch_off
         if not isinstance(envClass, type):
             raise Grid2OpException(
                 "Parameter \"envClass\" used to build the Runner should be a type (a class) and not an object "
@@ -470,7 +480,9 @@ class Runner(object):
                             opponent_action_class=self.opponent_action_class,
                             opponent_class=self.opponent_class,
                             opponent_init_budget=self.opponent_init_budget,
-                            name=self.name_env)
+                            name=self.name_env,
+                            ignore_min_up_down_times=self.ignore_min_up_down_times,
+                            forbid_dispatch_off=self.forbid_dispatch_off)
 
         if self.thermal_limit_a is not None:
             res.set_thermal_limit(self.thermal_limit_a)